Changed graphs to be smoother and began transfer of steering settings

+++ Chassi/Chassi.c
@@ -56,7 +56,7 @@ void arm_is_done(uint8_t id, uint16_t pickup_data)
 }
 
 uint16_t got_steering_request(uint8_t id, uint16_t metadata) {
-	return steering_wheel;
+	return (uint16_t)steering_wheel;
 }
 
 
@@ -140,6 +140,7 @@ int main(void)
 	bus_register_receive(12, engine_set_kd);
 	bus_register_receive(2, arm_is_done);
 	bus_register_response(5, got_steering_request);
+	
 	sei();
 	//enable timer interrupts for ocie0a
  	TIMSK0 |= (1 << OCIE0A);
